Added Tiles!
Added the possibility to use tiles as blocks, both in the game and in the editor. For reference use the image in the assets folder (tiles must be 16x16 and in that specific order)
Changed the block's dimensions to 32x32 (from 20x20) making previous made levels seamingly huge. The change was made to avoid non-integer scaling with the tiles. All the reference to the block dimension are now linked with the variable blockDimension, so that one change will make everything change accordingly.
Fixed the spikes that was manteining the old dimension (20x20) ruining the level aesthetic (only the editor had this bug).