Fix the visual connection of rooms in the map again
Connected rooms were overlapping each other in the map (2 blocks) now it's fixed, the script make the two connector-blocks make contact between each other without overlapping. If the block is the fartest of the map no overlap will occur, otherwise they will but that's how it should work. Level 2 is the perfect example, with this script the room with the leg armor is perfectly connected with the previous room, even if the connectorblock is not on the "rim" of the room. Instead both the entrance and the exit of the cave overlap with the connected room because the connector-block is not the last block. It's the map fault and not the code!